+++ tests/test_hdf5.py
@@ -1,5 +1,6 @@
 import os
 from types import SimpleNamespace
+from unittest.mock import patch
 
 import numpy
 import pytest
@@ -112,6 +113,42 @@ def example_file_with_links(tmp_path_factory):
         b["extr_link"] = h5py.ExternalLink("linked.h5", "/z/y")
 
     yield ensure_uri(file_path)
+
+
+@pytest.fixture(scope="module")
+def example_files_with_chunked_arrays(tmp_path_factory):
+    h5py = pytest.importorskip("h5py")
+    file_paths = []
+    for indx in range(3):
+        file_path = tmp_path_factory.mktemp("data").joinpath(
+            f"example_chunked_{indx}.h5"
+        )
+        with h5py.File(file_path, "w") as file:
+            a = file.create_group("a")
+            a.create_dataset(
+                "b",
+                data=numpy.arange(120 * indx, 120 * (indx + 1), dtype="int64").reshape(
+                    (4, 5, 6)
+                ),
+                chunks=(1, 2, 3),
+            )
+            a.create_dataset(
+                "c",
+                data=numpy.arange(10 * indx, 10 * (indx + 1), dtype="int64").reshape(
+                    (10, 1)
+                ),
+                chunks=(2, 1),
+            )
+            a.create_dataset(
+                "d",
+                data=numpy.arange(10 * indx, 10 * (indx + 1), dtype="int64").reshape(
+                    (10,)
+                ),
+                chunks=(3,),
+            )
+        file_paths.append(file_path)
+
+    yield [ensure_uri(file_path) for file_path in file_paths]
 
 
 def test_from_file(example_file, buffer):
@@ -571,3 +608,200 @@ def test_adapter_from_catalog(example_file, shape, error):
             adp.read()
     else:
         assert adp.read().shape == shape
+
+
+@pytest.mark.parametrize("num_files", [1, 3])
+def test_chunked_arrays_from_uris(example_files_with_chunked_arrays, num_files):
+    # Test that chunked arrays can be read and reshaped correctly
+    fpaths = example_files_with_chunked_arrays[:num_files]
+    tree = HDF5Adapter.from_uris(*fpaths)
+    with Context.from_app(build_app(tree)) as context:
+        client = from_context(context)
+
+    arr_b = client["a"]["b"]
+    assert arr_b.shape == (4 * num_files, 5, 6)
+    assert arr_b.chunks == ((1, 1, 1, 1) * num_files, (2, 2, 1), (3, 3))
+    assert arr_b.dtype == "int64"
+    arr_true_b = numpy.concatenate(
+        [
+            numpy.arange(120 * indx, 120 * (indx + 1), dtype="int64").reshape((4, 5, 6))
+            for indx in range(num_files)
+        ],
+        axis=0,
+    )
+    numpy.testing.assert_array_equal(arr_b.read(), arr_true_b)
+
+    arr_c = client["a"]["c"]
+    assert arr_c.shape == (10 * num_files, 1)
+    assert arr_c.chunks == ((2, 2, 2, 2, 2) * num_files, (1,))
+    assert arr_c.dtype == "int64"
+    arr_true_c = numpy.concatenate(
+        [
+            numpy.arange(10 * indx, 10 * (indx + 1), dtype="int64").reshape((10, 1))
+            for indx in range(num_files)
+        ],
+        axis=0,
+    )
+    numpy.testing.assert_array_equal(arr_c.read(), arr_true_c)
+
+    arr_d = client["a"]["d"]
+    assert arr_d.shape == (10 * num_files,)
+    assert arr_d.chunks == ((3, 3, 3, 1) * num_files,)
+    assert arr_d.dtype == "int64"
+    arr_true_d = numpy.concatenate(
+        [
+            numpy.arange(10 * indx, 10 * (indx + 1), dtype="int64").reshape((10,))
+            for indx in range(num_files)
+        ],
+        axis=0,
+    )
+    numpy.testing.assert_array_equal(arr_d.read(), arr_true_d)
+
+
+@pytest.mark.parametrize("num_files", [1, 3])
+@pytest.mark.parametrize("reshape", [True, False])
+def test_chunked_arrays_from_catalog(
+    example_files_with_chunked_arrays, num_files, reshape
+):
+    # Test that multiple chunked arrays can be read and reshaped correctly when initialized from catalog
+    assets = []
+    for indx, fpath in enumerate(example_files_with_chunked_arrays[:num_files]):
+        assets.append(
+            Asset(
+                data_uri=fpath,
+                is_directory=False,
+                parameter="data_uris",
+                num=indx,
+            )
+        )
+
+    shape = (4 * num_files, 5, 6) if not reshape else (num_files, 4, 5, 6)
+    chunks = (
+        ((1, 1, 1, 1) * num_files, (2, 2, 1), (3, 3))
+        if not reshape
+        else ((1,) * num_files, (1, 1, 1, 1), (2, 2, 1), (3, 3))
+    )
+
+    data_source = DataSource(
+        mimetype="application/x-hdf5",
+        assets=assets,
+        structure_family=StructureFamily.array,
+        structure=ArrayStructure(
+            shape=shape,
+            chunks=chunks,
+            data_type=BuiltinDtype.from_numpy_dtype(numpy.dtype("int64")),
+        ),
+        parameters={"dataset": "a/b"},
+        management=Management.external,
+    )
+
+    empty_node = SimpleNamespace(metadata_={}, specs=[])
+    adp = HDF5Adapter.from_catalog(data_source, empty_node, **data_source.parameters)
+    structure = adp.structure()
+    assert structure.shape == shape
+    assert structure.chunks == chunks
+    assert structure.data_type.to_numpy_dtype() == numpy.dtype("int64")
+
+    arr_true = numpy.concatenate(
+        [
+            numpy.arange(120 * indx, 120 * (indx + 1), dtype="int64").reshape((4, 5, 6))
+            for indx in range(num_files)
+        ],
+        axis=0,
+    )
+    if reshape:
+        arr_true = arr_true.reshape((num_files, 4, 5, 6))
+    numpy.testing.assert_array_equal(adp.read(), arr_true)
+
+
+@pytest.mark.parametrize("swmr", [True, False])
+def test_files_opened_and_closed(example_files_with_chunked_arrays, swmr):
+    "Test that only the necessary files are opened and that they are closed after reading"
+    import tiled
+
+    h5py = pytest.importorskip("h5py")
+
+    # Use the example with two files chunked along a single dimension;
+    # total chunks across the two files: ((3, 3, 3, 1)*2, )
+    file_uris = example_files_with_chunked_arrays[:2]
+    file_paths = [path_from_uri(uri) for uri in file_uris]
+    with patch(
+        "tiled.adapters.hdf5.h5open", wraps=tiled.adapters.hdf5.h5open
+    ) as mock_h5open:
+        mock_h5open.assert_not_called()  # No files should be opened yet
+
+        # Tree initialized from the entire file, no dataset provided
+        tree = HDF5Adapter.from_uris(*file_uris, swmr=swmr)
+        mock_h5open.assert_called()
+        assert mock_h5open.call_count == 1
+
+        # Adapter initialized directly from the dataset:
+        # Each file is read once to get the structure, and the first one
+        # is also read once again to get the metadata
+        mock_h5open.reset_mock()
+        HDF5ArrayAdapter.from_uris(*file_uris, dataset="a/d", swmr=swmr)
+        assert mock_h5open.call_count == 3
+        files_opened = [call.args[0].name for call in mock_h5open.call_args_list]
+        assert files_opened.count(file_paths[0].name) == 2
+        assert files_opened.count(file_paths[1].name) == 1
+        assert set(files_opened) == set([fp.name for fp in file_paths])
+
+        # Build the app fromn the tree (HDF5Adapter): all datasets in the file are parsed
+        mock_h5open.reset_mock()
+        with Context.from_app(build_app(tree)) as context:
+            client = from_context(context)
+        mock_h5open.assert_called()
+
+        # Initialize the array client -- no need to reopen the files
+        mock_h5open.reset_mock()
+        arr = client["a"]["d"]
+        assert arr.structure().shape == (20,)
+        assert arr.structure().chunks == ((3, 3, 3, 1) * 2,)
+        assert arr.metadata is not None
+        mock_h5open.assert_not_called()
+
+        # Read the entire array: files are opened once to get the specs and then again,
+        # four times each, to fetch each chunk separately. Additionally, the first file is opened once
+        # adain when initialized from catalog to get the metadata
+        assert arr.read() is not None
+        # 2 for specs, 4*2 for chunks, 1 for metadata
+        assert mock_h5open.call_count == 2 + 4 * 2 + 1
+        files_opened = [call.args[0].name for call in mock_h5open.call_args_list]
+        # First file: 1 for specs, 4 for chunks, 1 for metadata
+        assert files_opened.count(file_paths[0].name) == 4 + 1 + 1
+        # Second file: 1 for specs, 4 for chunks
+        assert files_opened.count(file_paths[1].name) == 4 + 1
+
+        # Read a slice that only touches one file: only the relevant file should be opened
+        mock_h5open.reset_mock()
+        assert arr[:1] is not None
+        assert mock_h5open.call_count == 4  # 2 for specs, 1 for chunks, 1 for metadata
+        files_opened = [call.args[0].name for call in mock_h5open.call_args_list]
+        assert files_opened.count(file_paths[0].name) == 3
+        assert files_opened.count(file_paths[1].name) == 1  # only to get the specs
+
+        # Read everything from the second file
+        mock_h5open.reset_mock()
+        assert arr[-10:] is not None
+        assert mock_h5open.call_count == 7
+        files_opened = [call.args[0].name for call in mock_h5open.call_args_list]
+        # First file: 1 for specs, 1 for metadata
+        assert files_opened.count(file_paths[0].name) == 1 + 1
+        # Second file: 4 for chunks, 1 for specs
+        assert files_opened.count(file_paths[1].name) == 4 + 1
+
+        # Read a slice that has one value from each of the files
+        mock_h5open.reset_mock()
+        assert arr[9:11] is not None
+        assert mock_h5open.call_count == 2 + 2 + 1
+        files_opened = [call.args[0].name for call in mock_h5open.call_args_list]
+        # First file: # 1 for specs, 1 for chunk, 1 for metadata
+        assert files_opened.count(file_paths[0].name) == 3
+        assert files_opened.count(file_paths[1].name) == 2  # 1 for specs, 1 for chunk
+
+    # Try opening the files directly to check that they are closed after reading
+    h5py.File(file_paths[0], "r", swmr=swmr).close()
+    h5py.File(file_paths[1], "r", swmr=swmr).close()
+
+    h5py.File(file_paths[0], "r", swmr=not swmr).close()
+    h5py.File(file_paths[1], "r", swmr=not swmr).close()